+++ src/main/java/com/bokmcdok/butterflies/event/entity/living/MobSpawnEventListener.java
@@ -2,7 +2,7 @@ package com.bokmcdok.butterflies.event.entity.living;
 
 import com.bokmcdok.butterflies.config.ButterfliesConfig;
 import com.bokmcdok.butterflies.registries.EntityTypeRegistry;
-import com.bokmcdok.butterflies.registries.TagRegistry;
+import com.bokmcdok.butterflies.world.entity.PeacemakerEntity;
 import com.bokmcdok.butterflies.world.entity.monster.PeacemakerButterfly;
 import net.minecraft.world.entity.*;
 import net.minecraft.world.entity.animal.*;
@@ -23,23 +23,14 @@ import net.minecraftforge.eventbus.api.IEventBus;
  */
 public class MobSpawnEventListener {
 
-    // The entity type registry.
-    private final EntityTypeRegistry entityTypeRegistry;
-    private final TagRegistry tagRegistry;
-
     /**
      * Construction
      * @param forgeEventBus The event bus to register with.
      */
-    public MobSpawnEventListener(IEventBus forgeEventBus,
-                                 EntityTypeRegistry entityTypeRegistry,
-                                 TagRegistry tagRegistry) {
+    public MobSpawnEventListener(IEventBus forgeEventBus) {
         forgeEventBus.register(this);
         forgeEventBus.addListener(this::onMobSpawn);
         forgeEventBus.addListener(this::onLivingDrops);
-
-        this.entityTypeRegistry = entityTypeRegistry;
-        this.tagRegistry = tagRegistry;
     }
 
     /**
@@ -65,7 +56,7 @@ public class MobSpawnEventListener {
             if (ironGolem.getRandom().nextInt() % 256 == 1) {
                 LevelAccessor levelAccessor = event.getLevel();
                 if (levelAccessor instanceof ServerLevelAccessor level) {
-                    EntityType<IronGolem> entityType = entityTypeRegistry.getButterflyGolem().get();
+                    EntityType<IronGolem> entityType = EntityTypeRegistry.BUTTERFLY_GOLEM.get();
 
                     if (ForgeEventFactory.canLivingConvert(ironGolem, entityType, (x) -> {
                     })) {
@@ -102,7 +93,7 @@ public class MobSpawnEventListener {
 
         // Don't infest entities if they are already infested.
         Entity entity = event.getEntity();
-        if (entity.getType().is(this.tagRegistry.getPeacemakerEntities())) {
+        if (entity instanceof PeacemakerEntity) {
             return;
         }