+++ src/main/java/com/bokmcdok/butterflies/registries/BannerPatternRegistry.java
@@ -3,29 +3,24 @@ package com.bokmcdok.butterflies.registries;
 import com.bokmcdok.butterflies.ButterfliesMod;
 import net.minecraft.core.Registry;
 import net.minecraft.world.level.block.entity.BannerPattern;
-import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.RegistryObject;
 
+/**
+ * Registers Banner Patterns used by the mod.
+ */
 public class BannerPatternRegistry {
 
-    // An instance of a deferred registry we use to register items.
-    private final DeferredRegister<BannerPattern> deferredRegister;
+    public static final DeferredRegister<BannerPattern> BANNER_PATTERNS;
+    public static final RegistryObject<BannerPattern> BANNER_PATTERN_BUTTERFLY;
 
-
-    /**
-     * Construction
-     * @param modEventBus The event bus to register with.
-     */
-    public BannerPatternRegistry(IEventBus modEventBus) {
-        this.deferredRegister = DeferredRegister.create(Registry.BANNER_PATTERN_REGISTRY, ButterfliesMod.MOD_ID);
-        this.deferredRegister.register(modEventBus);
+    static {
+        BANNER_PATTERNS = DeferredRegister.create(Registry.BANNER_PATTERN_REGISTRY, ButterfliesMod.MOD_ID);
+        BANNER_PATTERN_BUTTERFLY = BANNER_PATTERNS.register("banner_pattern_butterfly", () -> new BannerPattern("banner_pattern_butterfly"));
     }
 
     /**
-     * Register the banner patterns.
+     * Prevent construction.
      */
-    public void initialise() {
-        // Register the banner pattern itself.
-        deferredRegister.register("banner_pattern_butterfly", () -> new BannerPattern("banner_pattern_butterfly"));
-    }
+    private BannerPatternRegistry() {}
 }