+++ src/main/java/com/bokmcdok/butterflies/world/block/ButterflyMicroscopeBlock.java
@@ -1,7 +1,6 @@
 package com.bokmcdok.butterflies.world.block;
 
 import com.bokmcdok.butterflies.registries.BlockRegistry;
-import com.bokmcdok.butterflies.registries.ItemRegistry;
 import com.bokmcdok.butterflies.registries.MenuTypeRegistry;
 import com.bokmcdok.butterflies.world.inventory.ButterflyMicroscopeMenu;
 import net.minecraft.core.BlockPos;
@@ -38,29 +37,19 @@ public class ButterflyMicroscopeBlock extends Block {
     // The shape of the block.
     private static final VoxelShape SHAPE;
 
-    // Access to items.
-    private final ItemRegistry itemRegistry;
-
-    // Access to menu types.
-    private final MenuTypeRegistry menuTypeRegistry;
-
     /**
      * Construction
      */
-    public ButterflyMicroscopeBlock(ItemRegistry itemRegistry,
-                                    MenuTypeRegistry menuTypeRegistry) {
-        super(BlockBehaviour.Properties.of(Material.STONE)
-                .color(MaterialColor.STONE)
-                .isRedstoneConductor(BlockRegistry::never)
-                .isSuffocating(BlockRegistry::never)
-                .isValidSpawn(BlockRegistry::never)
-                .isViewBlocking(BlockRegistry::never)
+    public ButterflyMicroscopeBlock() {
+        super(BlockBehaviour.Properties.of()
+                .mapColor(MapColor.STONE)
+                .isRedstoneConductor(BlockRegistry::alwaysFalse)
+                .isSuffocating(BlockRegistry::alwaysFalse)
+                .isValidSpawn(BlockRegistry::alwaysFalse)
+                .isViewBlocking(BlockRegistry::alwaysFalse)
                 .noOcclusion()
                 .sound(SoundType.STONE)
                 .strength(1.0F));
-
-        this.itemRegistry = itemRegistry;
-        this.menuTypeRegistry = menuTypeRegistry;
     }
 
     /**
@@ -124,8 +113,7 @@ public class ButterflyMicroscopeBlock extends Block {
                                        inventory,
                                        title) ->
                 new ButterflyMicroscopeMenu(
-                        itemRegistry,
-                        menuTypeRegistry.getButterflyMicroscopeMenu().get(),
+                        MenuTypeRegistry.BUTTERFLY_MICROSCOPE_MENU.get(),
                         containerId,
                         inventory,
                         ContainerLevelAccess.create(level, blockPos)),