+++ src/main/java/com/bokmcdok/butterflies/world/entity/animal/Butterfly.java
@@ -2,7 +2,6 @@ package com.bokmcdok.butterflies.world.entity.animal;
 
 import com.bokmcdok.butterflies.ButterfliesMod;
 import com.bokmcdok.butterflies.config.ButterfliesConfig;
-import com.bokmcdok.butterflies.registries.BlockRegistry;
 import com.bokmcdok.butterflies.world.ButterflyData;
 import com.bokmcdok.butterflies.world.ButterflyInfo;
 import com.bokmcdok.butterflies.world.entity.DebugInfoSupplier;
@@ -55,7 +54,7 @@ import java.util.function.Predicate;
 
 /**
  * The butterfly entity that flies around the world, adding some ambience and
- * fertilising plants.
+ * fertilizing plants.
  */
 public class Butterfly extends Animal implements DebugInfoSupplier {
 
@@ -92,9 +91,6 @@ public class Butterfly extends Animal implements DebugInfoSupplier {
 
     // The location of the texture that the renderer should use.
     private final ResourceLocation texture;
-
-    // A reference to the block registry.
-    private final BlockRegistry blockRegistry;
 
     /**
      * Checks custom rules to determine if the entity can spawn.
@@ -222,12 +218,9 @@ public class Butterfly extends Animal implements DebugInfoSupplier {
      * @param entityType The type of the entity.
      * @param level The level where the entity exists.
      */
-    public Butterfly(BlockRegistry blockRegistry,
-                     EntityType<? extends Butterfly> entityType,
+    public Butterfly(EntityType<? extends Butterfly> entityType,
                      Level level) {
         super(entityType, level);
-
-        this.blockRegistry = blockRegistry;
 
         this.moveControl = new FlyingMoveControl(this, 20, true);
         this.setNoGravity(true);
@@ -313,14 +306,6 @@ public class Butterfly extends Animal implements DebugInfoSupplier {
         setNumEggs(numEggs);
 
         return super.finalizeSpawn(levelAccessor, difficulty, spawnType, groupData, compoundTag);
-    }
-
-    /**
-     * Accessor for the block registry.
-     * @return The block registry.
-     */
-    public BlockRegistry getBlockRegistry() {
-        return blockRegistry;
     }
 
     /**